pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-daily-brief-agent"
version = "0.1.0"
description = "Daily AI intelligence collection and push agent"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "APScheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
  "openai>=1.51.0",
  "pydantic>=2.8.2",
  "PyYAML>=6.0.2",
  "python-dateutil>=2.9.0.post0",
  "feedparser>=6.0.11",
  "lark-oapi>=1.4.16",
  "websockets<14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
